Add files via upload
This is the official esx_policejob, ready for qs-inventory and qs-smartphone, you just need to configure the config, and all was ready to go! 

I don't give any support on this script, because i'm not his developer. If you have any  problems with the quasar implementation, you can open a new commit and i check fastly! 

Read the config, and all is done!